Day 6
Adjusted the more options on each media settings and removed it from showing for Desktop devices. Also, made the footer bar become side bar on these devices and instead of the more options button, made a form in order to add a new consult on the bottom of the page.